- refactored code to remove duplicates
